+++ tests/test_snmp_data_client.py
@@ -48,7 +48,6 @@ class SnmpDataClientTestCase(unittest.IsolatedAsyncioTestCase):
                 device_type=device_type,
                 snmp_community="public",
                 poll_interval=0.1,
-                location="UnitTest",
             )
             snmp_data_client = epm.SnmpDataClient(
                 config=config, topics=topics, log=log, simulation_mode=1